+++ frontend/src/lib/components/AccountInfoForm.svelte
